requirements.txt
torch
numpy
opencv-python
matplotlib
scipy
pyyaml
pandas
pytest